chore: improve Tag exports
